GerarTextoServiceInterface: nova variável opcional. GerarTextoController: id do novo texto para realçar na view; nova variável para carregar na view e com retorno em json. GerarTextoService: nova variável, campos selecionados, retorno de objeto após atualizar, e evitar trazer conteúdo nos models do sumário. Migration create_gerar_texto: correção para tipo integer no campo ordem. custom.css: incluída classe de sortable dos textos. custom.js: alterações após mudança de layout. Blade textos: mudança de layout para facilitar manuseio de muitos textos. web: novo parâmetro opcional.

+++ database/migrations/2023_05_09_104907_create_gerar_textos.php
@@ -19,7 +19,7 @@ class CreateGerarTextos extends Migration
             $table->string('texto_tipo');
             $table->text('conteudo')->nullable();
             $table->boolean('com_numeracao')->default(true);
-            $table->unsignedTinyInteger('ordem')->default(1);
+            $table->unsignedInteger('ordem')->default(1);
             $table->string('nivel')->default(0);
             $table->string('tipo_doc')->default('carta-servicos');
             $table->string('indice')->nullable();
